num-squares-in-circle: Add initial version
Approach: Find the distance of each vertex from the center of the circle.
If all the vertices are lying within the two circles (ok if they are on
the circle itself), then count the square as a valid square.

TODO: The pattern of the squares within the circle would be symmetrical
across all quadrants. Hence we can only calculate for one quandrant and
then multiply by 4. Also we can compare only two vertices on extreme end.

Signed-off-by: Tushar Behera <trblinux@gmail.com>